Move Generators new directory
